added js doc and improved some names
